create MVVC+TableView for delegate and datasource
